Separate between code that is needed to test the storage, and code needed to create test storage to test other crates (externalized with "testing" feature)
