+++ crates/papyrus_storage/src/lib.rs
@@ -59,6 +59,9 @@ mod serializers;
 pub mod state;
 mod version;
 
+#[cfg(test)]
+mod test_instances;
+
 #[cfg(any(feature = "testing", test))]
 #[path = "test_utils.rs"]
 #[doc(hidden)]
